Move commonly used attributes into utils
